Implement the functionality that connects the MyBooksView to the /books endpoint.
make a BookRequests file with multiple book requests that subclass BaseBookRequest and conforms to Requestable.
update the Book model to include a container model of books.
create MyBooksViewModel that subclasses API to act as a service for the MyBooksView and data source. It's the interface to save, update, delete and fetch all the books.
update BookDataEntryView to initiate calls to the view model.
make BookRow dynamic
update MyBooksView to read in a book array from MyBookViewModel
create an OverlayProgressView to display while waiting for a response from the server.