Milestone: Sample app is fully functional with all views connected to the corresponding API. TODO: Perhaps some refactoring to tighten it up and some more robust error handling.
create MultipartFormdata object as a container to create httpBody content for a URLRequest. Also contains helper functions to transform a swift object into a data object to be sent.
create FieldType enum to support the MultipartFormdata field generation.
update Requestable to include a MultipartFormdata. Set the httpBody if the formData is present. Default is set to nil.
update ImageRequests to include a request for an image header.
change all ViewModels to be environment objects so that they are not destroyed and recreated. Memory cost here, but it's a sample app. Purpose is to demo networking.
update AddPhotoView to make an upload request, OverlayProgressView
refactor MyImagesView to break out the fetchPhoto logic into its own private function of the same name.
refactor performRequest in APIViewModel to handle .POST request types.